bug(UIKIT-485,ui,Datepicker): переезд Reason в общий types
