+++ packages/ui/src/types.ts
@@ -1,1 +1,11 @@
+import { AutocompleteCloseReason } from '@mui/base/AutocompleteUnstyled/useAutocomplete';
+
 export type WithoutEmotionSpecific<Props> = Omit<Props, 'sx' | 'css'>;
+
+export type Reason =
+  | 'escapeKeyDown'
+  | 'backdropClick'
+  | 'toggleInput'
+  | 'blur'
+  | AutocompleteCloseReason
+  | null;
